[ty] Extract shared abstract-method analysis
